fix: Return provided character
Parser functions now return the provided character, if they didn't
already. Previously if a check caused the function to return the
character would not get added to the array. Essentially "skipping"
it. This is now resolved and the character is added to the array
when all checks fail.

The proper substring is now passed into the atrules.oneliners indexOf
check of the artule parser. In turn, the flags.nested flag is now
properly set.