refactor: Update gulpfile.js
- Re-implement branch switch fail-safe. New way directly watches
the .git folder to detect for a branch change in real-time. This
makes for less code and is a more efficient approach to previous.
- Implements a way to detect and track when Gulp is running. This
is important for the clean-files Gulp task. This task should only
be run when Gulp is not running as it can produce many changes and
therefore cause many Gulp watcher runs.